Refactors task and task_group definitions out of DAG file and into module.
